+++ src/moaa_prime/fusion/mode.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class FusionMode(str, Enum):
+    ORACLE_MAX = "oracle_max"
+    ENERGY = "energy"